Rename plugins/inspector to plugins/help
