+++ spyderlib/plugins/help.py
